+++ CHANGES.md
@@ -1,4 +1,5 @@
 #5.0.32 20??-??-??
+ - 2018-12-31 Added option to limit number of active agent accounts.
  - 2018-12-04 Fixed upload cache filename cleanup and filename cleanup optimization.
  - 2018-12-03 Fixed infinite loop on filename generation.